requirements.txt
boto3>=1.34.0
PyYAML>=6.0.1

